now the classes themselves are responsible for reading their objects
Previously, the ConsoleManager was responsible for reading class objects. But if there are too many classes that need to be stored, then the manager will become too big. And when new stored classes appear, the ConsoleManager will have to be changed every time. Therefore, classes now read their own objects.